requirements.txt
# Backend Dependencies for Scallop Dredge Behavior Analysis

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
numpy==1.26.2
geopandas==0.14.1
pyarrow==14.0.1
shapely==2.0.2

# HTTP and async
aiohttp==3.9.1
requests==2.31.0

# Scientific computing
scipy==1.11.4
scikit-learn==1.3.2

# File processing
openpyxl==3.1.2
tabula-py==2.9.0
PyPDF2==3.0.1

# Geospatial
pyproj==3.6.1

# Machine learning model persistence
joblib==1.3.2

# Logging and utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Data validation
marshmallow==3.20.1
